Magasin + affichage map
Ajout de la classe magasin avec la méthode calcul
Affichage de la première map

+++ Advanced Tactics/Advanced Tactics/Unit.cs
@@ -8,9 +8,10 @@ namespace Advanced_Tactics
 {
     public class Unit
     {
+        //CLASSE QUI NE SERT A RIEN DESORMAIS
         public enum Type { troupe, commando, tank }
         public enum Rang { pion, fou, cavalier, tour, dame, roi }
-        private float niveau;
+        private int niveau;
         private int pos_x;
         public int x_unit
         {
@@ -29,25 +30,25 @@ namespace Advanced_Tactics
 
         public static void unite(Type type, float niveau, Rang rg)
         {
-            float prix = 1;
-            float force = 1;
-            float pv = 1;
+            int prix = 1;
+            int force = 1;
+            int pv = 1;
             switch (type)
             {
                 case Type.troupe :
-                    prix = 100 + 100 * (float)Math.Pow(1.5f, niveau);
-                    force = 10 + 10 * (float)Math.Pow(1.2f, niveau);
-                    pv = 50 + 50 * (float)Math.Pow(1.2f, niveau);
+                    prix = 100 + 100 * (int)Math.Pow(1.5f, niveau);
+                    force = 10 + 10 * (int)Math.Pow(1.2f, niveau);
+                    pv = 50 + 50 * (int)Math.Pow(1.2f, niveau);
                     break;
                 case Type.commando :
-                    prix = 500 + 500 * (float)Math.Pow(1.5f, niveau);
-                    force = 30 + 30 * (float)Math.Pow(1.2f, niveau);
-                    pv = 100 + 100 * (float)Math.Pow(1.2f, niveau);
+                    prix = 500 + 500 * (int)Math.Pow(1.5f, niveau);
+                    force = 30 + 30 * (int)Math.Pow(1.2f, niveau);
+                    pv = 100 + 100 * (int)Math.Pow(1.2f, niveau);
                     break;
                 case Type.tank:
-                    prix = 1200 + 1200 * (float)Math.Pow(1.5f, niveau);
-                    force = 90 + 90 * (float)Math.Pow(1.2f, niveau);
-                    pv = 300 + 300 * (float)Math.Pow(1.2f, niveau);
+                    prix = 1200 + 1200 * (int)Math.Pow(1.5f, niveau);
+                    force = 90 + 90 * (int)Math.Pow(1.2f, niveau);
+                    pv = 300 + 300 * (int)Math.Pow(1.2f, niveau);
                     break;
             }
             switch (rg)
@@ -59,7 +60,7 @@ namespace Advanced_Tactics
                     prix = prix * 2;
                     break;
                 case Rang.cavalier:
-                    prix = prix * 1.5f;
+                    prix = (int)(prix * 1.5f);
                     break;
                 case Rang.tour:
                     prix = prix * 2;
